example/requirements.txt
Django>=1.11
djangorestframework
django-rest-fuzzysearch
psycopg2
